requirements.txt
setuptools~=68.2.2
SQLAlchemy~=2.0.23
argparse~=1.4.0
PyYAML~=6.0.1
schema~=0.7.5